requirements.txt
opencv-python>=4.5.0,<5.0.0
numpy>=1.24.0,<2.0.0
mediapipe>=0.10.14,<0.11.0
deepface>=0.0.93
tensorflow>=2.15.0,<2.18.0
tf-keras>=2.15.0,<2.18.0
websocket-client>=1.6.0
protobuf>=4.25.3,<5.0.0
PySide6>=6.5.0
